make Note available in markdoc
