+++ docs/next/components/mdx/MDXComponents.tsx
@@ -18,6 +18,7 @@ import Zoom from 'react-medium-image-zoom';
 import {useVersion} from '../../util/useVersion';
 import Icons from '../Icons';
 import Link from '../Link';
+import {Note, Warning} from '../markdoc/Callouts';
 
 import 'react-medium-image-zoom/dist/styles.css';
 import {RenderedDAG} from './RenderedDAG';
@@ -233,53 +234,6 @@ const LinkGridItem = ({title, href, children, tags = []}) => {
       </div>
     </div>
   );
-};
-
-const ADMONITION_STYLES = {
-  note: {
-    colors: {
-      bg: 'primary-100',
-      borderIcon: 'primary-500',
-      text: 'gray-900',
-    },
-    icon: Icons.About,
-  },
-  warning: {
-    colors: {bg: 'yellow-50', borderIcon: 'yellow-400', text: 'yellow-700'},
-    icon: Icons.About,
-  },
-};
-
-const Admonition = ({style, children}) => {
-  const {colors, icon} = ADMONITION_STYLES[style];
-  return (
-    <div className={`bg-${colors.bg} border-l-4 border-${colors.borderIcon} px-4 my-4`}>
-      <div className="flex items-center">
-        <div className="flex-shrink-0">
-          <svg
-            className={`h-5 w-5 text-${colors.borderIcon}`}
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 25 25"
-            fill="currentColor"
-            aria-hidden="true"
-          >
-            {icon && icon}
-          </svg>
-        </div>
-        <div className="ml-3">
-          <p className={`text-sm text-${colors.text}`}>{children}</p>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const Note = ({children}) => {
-  return <Admonition style="note">{children}</Admonition>;
-};
-
-const Warning = ({children}) => {
-  return <Admonition style="warning">{children}</Admonition>;
 };
 
 const CodeReferenceLink = ({filePath, isInline, children}) => {
