+++ docs/next/markdoc/tags.tsx
@@ -1,0 +1,9 @@
+import {Note, Warning} from '../components/markdoc/Callouts';
+
+export const note = {
+  render: Note,
+};
+
+export const warning = {
+  render: Warning,
+};
